Move handle_parent_notification logic into the newly created Child.LifecycleController. Use this logic in both bin and element implementation
