+++ lib/membrane/core/bin.ex
@@ -10,7 +10,6 @@ defmodule Membrane.Core.Bin do
   alias Membrane.Core.{
     CallbackHandler,
     Child,
-    Component,
     Message,
     Parent,
     Telemetry,
@@ -20,7 +19,6 @@ defmodule Membrane.Core.Bin do
   require Membrane.Core.Message
   require Membrane.Core.Telemetry
   require Membrane.Logger
-  require Membrane.Core.Component
 
   @type options_t :: %{
           name: atom,
@@ -142,16 +140,7 @@ defmodule Membrane.Core.Bin do
 
   @impl GenServer
   def handle_info(Message.new(:parent_notification, notification), state) do
-    context = Component.callback_context_generator([:Bin], ParentNotification, state)
-
-    state =
-      CallbackHandler.exec_and_handle_callback(
-        :handle_parent_notification,
-        Membrane.Core.Bin.ActionHandler,
-        %{context: context},
-        notification,
-        state
-      )
+    state = Child.LifecycleController.handle_parent_notification(notification, state)
 
     {:noreply, state}
   end
